Remove Segments module; functions moved to CommandQueue
