+++ lib/chukinas/dreadnought/command_queue.ex
@@ -1,4 +1,5 @@
-alias Chukinas.Dreadnought.{CommandQueue, Command, Segments}
+alias Chukinas.Dreadnought.{CommandQueue, Command, Segment}
+alias Chukinas.Geometry.Rect
 defmodule CommandQueue do
 
   # *** *******************************
@@ -39,6 +40,21 @@ defmodule CommandQueue do
   end
 
   def build_segments(command_queue, start_pose, arena) do
-    Segments.init(command_queue, start_pose, arena)
+    starts_inbounds? = get_inbounds_checker(arena)
+    command_queue
+    |> Stream.scan(start_pose, &Command.generate_segments/2)
+    |> Stream.take_while(starts_inbounds?)
+    |> Stream.concat()
+    |> Enum.to_list()
+  end
+
+  # *** *******************************
+  # *** PRIVATE
+
+  defp get_inbounds_checker(arena) do
+    fn segments ->
+      pose = segments |> List.first() |> Segment.get_start_pose()
+      arena |> Rect.contains?(pose)
+    end
   end
 end
